# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing mechanics, customers, inventory, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000" # render name: "your-app-name.onrender.com"
schemes:
  - "https"
basePath: "/"
securityDefinitions:
  BearerAuth:
    type: apiKey
    swagger: "2.0"
    in: header
    description: "JWT token using Bearer token number. Example: 'Bearer {token number}'"
paths:
  /inventory/mechanic/:
    get:
      tags: [Inventory]
      summary: "Mechanic gets paginated list of parts with pricing"
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          type: integer
        - name: per_page
          in: query
          type: integer
      responses:
        200:
          description: "List of parts with pricing"
        400:
          description: "Bad request"
        500:
          description: "Internal server error"
  /inventory/mechanic/{id}:
    get:
      tags: [Inventory]
      summary: "Mechanic gets part by id with pricing"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Part details with pricing"
        404:
          description: "Not found"
        500:
          description: "Internal server error"
  /inventory/:
    post:
      tags:
        - Inventory
      summary: Mechanic adds a new part
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: part
          required: true
          schema:
            $ref: '#/definitions/InventoryItem'
      responses:
        201:
          description: Part added
        400:
          description: Bad request
    get:
      tags:
        - Inventory
      summary: Public gets paginated list of parts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          required: false
        - name: per_page
          in: query
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: List of parts
        '400':
          description: Bad request
  /inventory/{id}:
    get:
      tags:
        - Inventory
      summary: Public gets part by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Part details
        '404':
          description: Not found
    put:
      tags:
        - Inventory
      summary: Mechanic updates part
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - in: body
          name: part
          required: true
          schema:
            $ref: '#/definitions/InventoryItem'
      responses:
        200:
          description: Part updated
        400:
          description: Bad request
    delete:
      tags:
        - Inventory
      summary: Mechanic deletes part
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Part deleted
        '400':
          description: Bad request
  /inventory/search:
    get:
      tags:
        - Inventory
      summary: Public searches parts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Search results
        '400':
          description: Bad request
  /inventory/mechanic/search:
    get:
      tags:
        - Inventory
      summary: Mechanic searches part by name
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search result
        '400':
          description: Bad request
        '404':
          description: Not found
  /inventory/low-stock:
    get:
      tags:
        - Inventory
      summary: Mechanic gets low stock parts
      security:
        - BearerAuth: []
      parameters:
        - name: threshold
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of low stock parts
        '400':
          description: Bad request
...
definitions:
...
  InventoryItem:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    required:
      - name
      - price
...